refactor: Mesh import + scale setter usage
