+++ public/hl2/Renderer.js
@@ -1,6 +1,7 @@
-import {Mesh, Scene} from "../../src/index.js";
+import {Scene} from "../../src/index.js";
 import {ShaderLoader} from "../../src/Loader/index.js";
 import {Material} from "../../src/Material/index.js";
+import {Mesh} from "../../src/Mesh/index.js";
 import {WebGLRenderer} from "../../src/Renderer/index.js";
 import {SSDPlaneGeometry} from "./SSDPlaneGeometry.js";
