+++ public/hl2/Scene.js
@@ -1,7 +1,8 @@
-import {Mesh, Scene as _Scene} from "../../src/index.js";
+import {Scene as _Scene} from "../../src/index.js";
 import {PointLight} from "../../src/Light/index.js";
 import {Material} from "../../src/Material/index.js";
 import {Matrix4, Vector3} from "../../src/math/index.js";
+import {Mesh} from "../../src/Mesh/index.js";
 
 export class Scene extends _Scene {
 	/**
