+++ public/minecraft/scenes/instancing.js
@@ -1,7 +1,8 @@
-import {Camera, Mesh} from "../../../src/index.js";
+import {Camera} from "../../../src/index.js";
 import {BoxGeometry} from "../../../src/Geometry/index.js";
 import {Material} from "../../../src/Material/index.js";
 import {Matrix3, PI, Vector2, Vector3} from "../../../src/math/index.js";
+import {Mesh} from "../../../src/Mesh/index.js";
 import {BLOCK_SCALE, FIELD_OF_VIEW} from "../main.js";
 import {Scene} from "../Scene.js";
 
@@ -30,7 +31,7 @@ export function createScene() {
 		if (i % 100 === 0) k++;
 
 		mesh.setPosition(new Vector3(i % 10 - 4.5, 1 - k, j % 10 - 4.5));
-		mesh.scale = new Vector3().addScalar(BLOCK_SCALE);
+		mesh.getScale().multiplyScalar(BLOCK_SCALE);
 
 		meshes.push(mesh);
 	}
