+++ public/visibility/scenes/visibility.js
@@ -1,6 +1,7 @@
-import {Camera, Mesh, Scene} from "../../../src/index.js";
+import {Camera, Scene} from "../../../src/index.js";
 import {BoxGeometry} from "../../../src/Geometry/index.js";
 import {PI, Vector2, Vector3} from "../../../src/math/index.js";
+import {Mesh} from "../../../src/Mesh/index.js";
 import {SENSITIVITY} from "../../hl2/main.js";
 import {CAMERA_HEIGHT, FIELD_OF_VIEW} from "../../minecraft/main.js";
