+++ src/Mesh/index.js
@@ -1,0 +1,2 @@
+export {Mesh} from "./Mesh.js";
+export {PlayerMesh} from "./PlayerMesh.js";
