+++ src/index.js
@@ -3,6 +3,5 @@ export {Camera} from "./Camera.js";
 export {Debugger} from "./Debugger.js";
 export {Hitbox} from "./Hitbox.js";
 export {Instance} from "./Instance.js";
-export {Mesh} from "./Mesh.js";
 export {Scene} from "./Scene.js";
 export {TextureImage} from "./TextureImage.js";
